Shift grade and attendance UI to only for verbose
